refactor: 입력과 유효성 검사 관련 메소드들은 Input.js의 Input 클래스로 분리
